Put tie line tests apart
